football_club/requirements.txt
django == 4.2.8
django-filter == 23.5
env python = 3.11.1